requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.7
pandas==2.1.3
